feat: add support for static export in shared collection and public profile pages
- Implemented `generateStaticParams` for shared collection and public profile pages to support Capacitor builds.
- Added placeholder routes to skip fetch and render loading states for static exports.
- Updated SEO metadata generation to handle static export cases.

feat: introduce client-side item and manifestation pages

- Created `item-page-client.tsx` to display item details with polling for updates.
- Developed `manifestation-page-client.tsx` to show manifestation details, including metadata and reviews.
- Integrated loading states and error handling for both item and manifestation pages.

fix: enhance locale handling in i18n request configuration

- Modified locale retrieval to only access cookies in non-export mode, improving compatibility with static exports.
